chore: move API service to libs
